+++ test/text_encoder_test.py
